+++ packages/react-ui/src/components/PaginationTwo/index.tsx
@@ -17,14 +17,6 @@ const Wrapper = styled.div<{ mobile?: boolean }>`
   flex: 1 1 auto;
   list-style: none;
   color: ${({ theme }) => theme.color['Neutral/Neutral 90']};
-
-  ${({ mobile }) =>
-    mobile &&
-    `
-  & > ul {
-    margin: 0 4px 0 4px;
-  }
-  `}
 `;
 const PagesWrapper = styled.ul`
   position: relative;
@@ -43,6 +35,7 @@ const Input = styled(TextInput)`
 `;
 
 const Button = styled(SecondaryButton)`
+  width: 100%;
   margin-bottom: 20px;
 `;
 
@@ -162,14 +155,13 @@ export const PaginationTwo: React.FC<PaginationTwoProps> = ({
 
   const onInputSubmit = (event: any, value: string) => {
     const parsed = parseInt(value, 10);
-    if (isNaN(parsed)) {
-      setInputValue('');
-    } else if (disabledPages.indexOf(parsed) > -1) {
-      setInputValue('');
+    setInputValue('');
+
+    if (isNaN(parsed) || disabledPages.indexOf(parsed) > -1) {
+      return;
     } else {
       if (parsed !== page) {
         const newValue = parsed > count ? count : parsed < 1 ? 1 : parsed;
-        setInputValue(newValue.toString());
         onChange?.(event, newValue);
       }
     }