+++ packages/react-ui/src/components/form/SuggestField/SuggestField.stories.tsx
@@ -1,4 +1,4 @@
-import { INPUT_DIMENSIONS_VALUES } from '#src/components/input';
+import { INPUT_DIMENSIONS_VALUES, INPUT_STATUS_VALUES } from '#src/components/input';
 import * as React from 'react';
 import type { ChangeEvent } from 'react';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
@@ -39,6 +39,10 @@ export default {
     dimension: {
       options: INPUT_DIMENSIONS_VALUES,
       control: { type: 'radio' },
+    },
+    status: {
+      control: { type: 'radio' },
+      options: INPUT_STATUS_VALUES,
     },
     isLoading: {
       control: { type: 'boolean' },